Come on Baby!
Well, we can say that this commit refers to 30% of the project.

This refers to the structuring of the site, such as homepage, bot, add bot and vote.

Everything is working well, several things are still incomplete such as the comment system and the voting system.

As I said at the beginning, it's a slow job and it will take a long time to complete!

That's it, Bye